[FIX] send correct timings
`Sender::run` sends the content of `data` in an endless loop.
A second thread has to call `Sender::send` to update `data`. In most
cases this happens more often than the loop being executed. In some rare
occasions this is not the case.
Since `Sender::run` is modifying `data` on each loop the values become
invalid, causing the whole function to send out wrong data.

This commit fixes this, by deleting `data` each time it is being send.